requirements.txt
pandas
streamlit
pyserial
scikit-learn
joblib
plotly
